Support recursive grammars by resetting CSS scopes
Under our new TextMate-based highlighting, we can end up with nested
highlighting <span>s. In some cases we need to reset the highlighting
styles to match the top-level styles. This makes it so that, e.g., code
inside interpolated strings is colored as code rather than as string
text. The new <span class="h"> elements that the TextMate highlighter
produces indicate where this needs to happen, and the styles added in
this commit accomplish the reset.

See https://github.com/github/lex-luthor/issues/31 and
https://github.com/github/github/pull/34312 for more discussion.